backend/requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
python-dotenv
pydantic
python-multipart
google-generativeai
pdf2image
Pillow
faster-whisper

